+++ domain/course/src/main/java/com/teamsolply/solply/course/favoriteTown/model/Town.kt
@@ -1,6 +1,6 @@
 package com.teamsolply.solply.course.favoriteTown.model
 
 data class Town(
-    val townId: Int,
+    val townId: Long,
     val townName: String
 )
