+++ domain/course/src/main/java/com/teamsolply/solply/course/favoriteTown/repository/FavoriteTownRepository.kt
@@ -5,7 +5,7 @@ import com.teamsolply.solply.course.favoriteTown.model.FavoriteTownInfoEntity
 interface FavoriteTownRepository {
     suspend fun getTownList(): Result<FavoriteTownInfoEntity>
     suspend fun patchUserFavoriteTown(
-        selectedTownId: Int,
-        favoriteTownIdList: List<Int>
+        selectedTownId: Long,
+        favoriteTownIdList: List<Long>
     ): Result<Unit>
 }
