+++ feature/course/src/main/java/com/teamsolply/solply/course/favoriteTown/model/CourseState.kt
@@ -1,6 +1,6 @@
 package com.teamsolply.solply.course.favoriteTown.model
 
 data class CourseState(
-    val selectedTownId: Int?,
+    val selectedTownId: Long?,
     val townList: List<TownModel>
 )
